with-apim/requirementes.txt
openai==1.61.1
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-ai-documentintelligence==1.0.0
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.36
langchain-openai==0.3.6
